[#2] Updated Vade.py to support Discord.py rewrite
