+++ src/VadeBot.py
@@ -9,7 +9,7 @@ from discord.ext.commands import Bot
 from src.util import db
 
 bot = Bot(description="FUCK THIS SHIT", command_prefix=os.environ["Command_Prefix"], pm_help=False)
-extensions = ['Utility', 'Mathematics']
+extensions = ['Utility', 'Mathematics', 'Vade']
 # extensions = ['Utility', 'Mathematics', 'Vade', 'NOHK', 'Rainbow6']
 
 file_list = glob.glob(os.path.join(os.getcwd(), "src/files/prompts", "*.txt"))
@@ -23,11 +23,6 @@ for path in file_list:
         for line in lines:
             messages.append(line)
 
-pic_list = []
-for file in os.listdir("src/pics"):
-    if file.lower().endswith(".png") or file.lower().endswith(".jpg"):
-        pic_list.append('src/pics/{}'.format(file))
-
 card_list = []
 for file in os.listdir("src/cards"):
     if file.lower().endswith(".png") or file.lower().endswith(".jpg"):
@@ -35,15 +30,8 @@ for file in os.listdir("src/cards"):
 card_map = [x.replace("src/cards/", "").replace(".png",
                                                "").replace(".jpg", "").lower() for x in card_list]
 
-ball_replies = []
-with open("src/files/8ballReplies.txt") as file:
-    ball_replies = [line.strip() for line in file]
-
-#mudae_ids = ["432610292342587392", "588992629136424960", "531931447129538588"]
-
 def bobo_tag(user_id):
     return 'BOBO MO <@{!s}>'.format(user_id)
-
 
 def find_Bobo(words):
     return 'bobo' in words
@@ -63,18 +51,7 @@ async def on_message(message):
     user_id = message.author.id
     print(message)
     words = message.content.lower().split()
-    
-    #if (user_id in mudae_ids and message.embeds):
-    #    await bot.add_reaction(message, u"\u2B05")
-    #    await bot.add_reaction(message, u"\u27A1")
-    #    await asyncio.sleep(3)
-    #    await bot.add_reaction(message, 	u"\U0001F496")
     if user_id != bot.user.id and message.guild.id != 607181202922799135:
-        # if message.content.startswith('v!8ball'):
-        #     if len(words) == 1:
-        #         await message.channel.send("THAT AIN'T A FUCKING QUESTION FFS")
-        #     else:
-        #         await message.channel.send(random.choice(ball_replies))
         if message.content.lower() == "good vade":
             await message.channel.send(file=discord.File('src/pics/vadesmile.jpg'))
         elif message.content.lower() == "bad vade":
